changed int for size_t

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -8,7 +8,7 @@
 
 size_t print_listint(const listint_t *h)
 {
-	int nnode;
+	size_t nnode;
 
 	for (nnode = 0; h != NULL; nnode++)
 	{
